ReviewerにBadgeを追加
